fix: hanging lambdas
For some reason some of my API Gateways and Kinesis Lambdas return a success object but the lambda hangs and never completes until it's reached the maximum time limit for the Lambda. When I remove the override of the context handler (moving the JS transform into async-lambda-fn) it never times out.